change the design of main screen and display note screen , edit the content of note directly

+++ lib/widgets/note_item.dart
@@ -50,25 +50,56 @@ class NoteItem extends StatelessWidget {
             },
             title: Text(
               note.title,
+              maxLines: 2,
               style: const TextStyle(
                 color: Colors.black,
                 fontSize: 26,
+                overflow: TextOverflow.ellipsis,
               ),
             ),
             subtitle: Padding(
               padding: const EdgeInsets.only(top: 16, bottom: 12),
               child: Text(
                 note.subtitle,
+                maxLines: 2,
                 style: TextStyle(
                   color: Colors.black.withOpacity(0.5),
                   fontSize: 18,
+                  overflow: TextOverflow.ellipsis,
                 ),
               ),
             ),
             trailing: IconButton(
               onPressed: () {
-                note.delete();
-                BlocProvider.of<NotesCubit>(context).fetchAllNotes();
+                showDialog(
+                    context: context,
+                    builder: (context) => AlertDialog(
+                          title: const Text(
+                              'Are you sure you want to delete this note ?'),
+                          actions: [
+                            TextButton(
+                              onPressed: () {
+                                note.delete();
+                                BlocProvider.of<NotesCubit>(context)
+                                    .fetchAllNotes();
+                                Navigator.pop(context);
+                              },
+                              child: const Text(
+                                "Delete",
+                                style: TextStyle(color: Colors.red),
+                              ),
+                            ),
+                            TextButton(
+                              onPressed: () {
+                                Navigator.pop(context);
+                              },
+                              child: const Text(
+                                "Cancel",
+                                style: TextStyle(color: Colors.blue),
+                              ),
+                            ),
+                          ],
+                        ));
               },
               icon: const Icon(
                 FontAwesomeIcons.trash,
